metrics_module/metrics_api/requirements.txt
aiokafka==0.10.0
async_fastapi_jwt_auth==0.6.2
backoff==2.2.1
fastapi==0.100.1
gunicorn==21.2.0
uvicorn==0.24.0
pydantic_settings==2.1.0
pydantic[email]==2.5.2
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-exporter-jaeger==1.21.0
pyjwt==2.8.0
sentry-sdk[fastapi]==1.40.6
python-logstash==0.4.8
python-logstash-async==3.0.0

